requirements.txt
# Email Scraper Pro v3.0 - Requirements
# Install with: pip3 install -r requirements.txt

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# DNS resolution for email validation
dnspython>=2.4.0

# SOCKS proxy support (for Tor)
PySocks>=1.7.1

# Optional: For colored output on Windows
colorama>=0.4.6
